Use feature-agnostic endpoints.
